+++ src/options-validation/encrypt-props.js
@@ -1,7 +1,10 @@
+// Validates encryption option properties.
+
 const validationTasks = require("../common/validation-tasks");
 const numberLimits = require("../common/number-limits");
 
 
+// Main function.
 function validateEncryptionProperties(optionsObject, resultObject)
 {
 	handleTrueFalseValue(optionsObject.encryption, "enabled", resultObject);
@@ -10,15 +13,7 @@ function validateEncryptionProperties(optionsObject, resultObject)
 }
 
 
-function handleEnabledStatus(givenValue, resObject)
-{
-	if (resObject.valid === true)
-	{
-		validationTasks.checkBoolean(givenValue, "encryption.enabled", resObject);
-	}
-}
-
-
+// Boolean value.
 function handleTrueFalseValue(encryptObj, propName, resObject)
 {
 	var givenValue = encryptObj[propName];
@@ -31,6 +26,7 @@ function handleTrueFalseValue(encryptObj, propName, resObject)
 }
 
 
+// Key string.
 function handleKeyString(givenString, resObject)
 {
 	if (resObject.valid === true)